Move data.json export code
Move the export code so that it can be shared between the pylons controller and
the flask controller.
